Test drive code math with Desmos and HTML demo
Made a quick demo to see if the drive code is correct or not. driveCodeTest.html also has a link to the first proof in Desmos.
The tracking math assumes that the drive math is correct, but I never actually proved if it's correct. Now I know it is correct, so the tracking issues must be from something else!