Add InertialIntercept class to expose the vex::inertial::temperature() function
Created a class in dualInertial.h to replace the vex::inertial class. This new class has a public GetTemperature() method.
In the few 18000 degree tests I've done so far, it seems like the results can vary. I wonder if this is a result of the sensors heating up over time. This intercept class will allow me to track temperature in my tests.