Test first prototype of APS paths
It's now possible to add a list of points, then the robot will repeatedly try driving to the first one and removing it within 5 inches. Many bugs probably, everything is hardcoded, numbers need to be tweaked and fine tuned still.
The program is not in a good state, but this is the first version that could drive a few laps automatically.