Remove automatic backwards driving in APS
Removed a line of code that would make the robot align with the opposite direction if the target was behind the robot.
Backwards driving was added to stop the robot from rotating a lot once it gets super close to the target, but it skips targets now. In a path, it was causing many left turns to turn excessively.