Add APS::EndPath()
Added a function to wait until the path is finished driving. Bound autonomous to the A button and made it so usercontrol is called once autonomous is finished.
Other motors like the lift will need to be called in the middle of the drivetrain commands. Therefore, the paths are only meant to buffer a small amount, drive that small amount, do something else elsewhere, then buffer another section of path.