Add intakeAlignButton
Added a bumper button that will aid in intake alignment. Hold it to rotate the intake, fine adjust manually, then press again to reset encoder.
This will make it easier and quicker to set up the robot, since the intake hook no longer has to be moved long distances with the tiny gear.