Add options to the PathPoint struct
Added the ability to set driveBackwards, maxDriveSpeed, maxTurnSpeed, and distanceThreshold for each PathPoint.
This will make it possible to fine tune the exact arc the robot drives getting to each point. Sometimes it can make a wide arc and drive quickly, but other times it has to turn in a tight space.