Remove new intake system and add new claw
Removed the intake subsystem files and disabled the MoveStoredRingTo___ controls. Renamed current claw code to clawPivot and added a new claw system.
The hook intake from Lobstah Bowl works well, so we're reverting back to that for York. The claw needs to be redesigned for picking things up off the ground too.