Add reset button
Set the A button to trigger a reset routine on the lift and clawPivot. This routine will lower the torque and move the motors as far against another part of the robot as possible, then reset the encoders to zero.
Sometimes, I think when the brain draws too much power or something, out lift mechanisms suddenly power off for a second. If alignment between the lift and clawPivot is to be synced, there needs to be a button to manually sync them again in the event of this power issue.