Bind clawPivot controls to L1 and L2 like lift
Set the clawPivot to use the same buttons as the lift, making it automatically go to an optimum angle when grabbing or placing a ring.
Having to line up two different systems to put a ring onto the wall stakes is work that can be reduced. Controlling both at the same time will make it faster and more consistent.