Add manual controls back to clawPivot
Re-added the the R1 and R2 manual controls for the clawPivot motor. It will still be controlled automatically, but manual control can be taken if needed.
Manual control helps adjust the angle better, in case trying to put the ring onto a mobile goal or something from the claw which is a different height.