Combine lift and clawPivot into clawArm
Copied the code for the two separate subsystems into one subsystem. Also renamed the lift to the arm to match the rest of the notebook better.
Both of these subsystems do very similar things, might as well make them the same. Refactoring the make a combined control setup is also the next todo, so this is step one.