Rewrite clawArm subsystem to stop over expansion
Holding the up button will now raise the clawArm in three different segments. First turning the claw, then the arm, then letting the claw back out as the arm goes up higher.
This stops the arm and claw from extending past expansion limits. The robot is over 24 inches wide when the arm is extended to its fullest, but this code prohibits that from happening.