End of York Competition
This commit holds all of the revisions done during the Nov 16 York competition. Some things include resurrecting and tweaking Lobstah Bowl skills auton, making positive side auton move off the starting line, moving hooks, adjusting the clawArm angles, and moving a port.
Lobstah Bowl auton was easy to uncomment and adjust a few names, making it a good candidate for getting a score up. Hooks were moved for a variety of reasons, such as removing and adding chain links or needing to reassemble the entire belt after a large snap. The claw was pointing a little too upward at its highest position, making it difficult to get rings onto the neutral wall stake. The brain was rotated to stop the mogoMover wire from getting caught on the hooks, and a single port was moved to aid in that rotation.