Add clawLift subsystem
Write some basic code to move the clawLift up and down and toggle the two pistons.
This code will come in handy once both sides of the lift are built, since the MotorTesting program can only test one motor at a time. It will likely also be the desired behavior for competition.