Fix clawLift direction
Flipped the direction of the clawLift motors in robotConfig.cpp so that holding L1 will move the claw up instead of down.
L1 should always be bound to the forward direction since "move up" logically matches with forward. The only other way to switch a motor's direction is in robotConfig.cpp.