Switch toggle mogo mover and toggle claw buttons
Set B to be the mogo mover and Y to be the claw. Few other minor changes like updating notes for R1 and R2, which were modified at York without updating notes.
Emma's request. She was mixing them up at the York competition and thought it'd be good to switch them.