Re-enable autonomous code with the clawArm code removed
Uncomment and slightly clean up the York autonomous code. Removed the code that would raise the arm so there aren't any compiler errors about the now undefined motors.
This code will likely be reused for the Yarmouth 1 competition. It can be uncommented for syntax highlighting and pasted into the programming notebook.