Add manual rightClawLift controls to axis 2
Made it so pushing axis 2 all the way up or all way down will manually move the right side of the clawLift.
If the two sides of the clawLift become unaligned, it becomes very difficult for the linear slides to move along the track. Being able to move one motor at a time will allow us to realign the motors in this scenario.