APC Add vertex UI in a side panel
Added a side panel that can precisely control the coordinate of path points with a keyboard instead of the mouse.
This side panel will be needed to add extra code to the points. It was originally going to be one large textedit tag to modify the JSON of each point, but it looked ugly so I'm creating separate UI elements for each value.