APC Draw field as the background image
Added two pictures for the match and skills field from the game manual, and a UI dropdown to select between which one is drawn. Also fixed a few UI bugs.
Being able to see the field behind the path will make it much easier to place the points on different objects.