Prototype neutral flip and elevation subsystems
Wrote quick prototypes for the elevation subsystem and the new subsystem to put rings onto the neutral stake. Also switched a few ports around and added to the controller monitors.
Emma and Ava built these systems over the past week, and now we're at the Coliseum to test them out. They need a quick program so we can start testing as soon as possible.