APC Add controls for splitting, removing, and prepending to the path
Arrow keys can now add points to the path in three different ways, and the backspace key will delete the selected point. Controls are now printed in the console, and fixed an bug with JSON input not setting a point's code.
These new keyboard controls will make it much easier and quicker to make some changes to the path.