APC Bug fixes
Removed an extra parenthesis in the exported C++ code. Made sure to append "aps.EndPath();" with or without a code point. Splitting or removing PathSections will keep code point in the same spot.
Having the code points stay in the same spot while splitting or removing PathSections is usually the preferred behavior.