Update APS to accept the APC's Bezier curves and add GPS
APS now calculates the next point instead of simply retrieving it. Attempts to use the GPS sensor to help correct position and show results on screen.
Using a smooth curve to get the target point gives more control and increased accuracy. A GPS sensor will theoretically help with correcting for the APS's drift.