Add unpowered tracking wheel to APS
Added a shaft encoder and modified the APS to calculate horizontal movement.
This will help improve tracking quality, especially when the robot tries to move quickly across the field, because the shaft encoder can detect the slipping and sliding motion that is perpendicular to the drivetrain wheels. The math may not be exact, but the compensation it provides does seem to help more than it impedes. The tracking wheel was already built and previously attempted to implement, but the code was only now written because the unpowered wheel mechanism was only just now made slim enough to fit behind the mogo mover without hitting the mogo.