Remove debug code & disable GPS
Removed many print statements and debug code that was strewn throughout the code to help with debugging the APS. Commented out the code that uses the GPS's data, although the sensor is still in robotConfig.cpp.
Everything now works as intended, the console clutter and debug code can be removed. GPS does not work as intended, but if turns out that it's more accurate on an official field, it may still be useful.