Retract neutral flip at end of autonomous and driver control
Added an event to call RetractNeutralFlip 5 seconds from the end of match driver control, and call it near the end of autonomous sequences.
If the neutral flip mechanism is extended while power is cut, it will fall down, go out of size, and potentially DQ us. A physical stop was going to be added, but, we didn't have time.