Hit the ladder with the front of the robot in auton
Updated autonomous to turn around and drive into the ladder with the front of the robot instead of the back. Also changed excentuated a curve so the robot drives into the ring with a better line.
Backing up into the ladder makes the mogo go under the ladder, potentially getting stuck. Turning around makes the start of driver control easier and faster.