Retract neutral flip during autonomous with an event
Set up an event to retract the neutral flip mechanism at the end of autonomous in the same way that it does in driver control.
If the robot gets stuck, it fails to reach the RetractNeutralFlip() function in the autonomous path. This trigger based on time instead of progression through the path ensures it always gets called.