Replace elevation pistons with motors
Switched elevationPistons to elevationArm, a motor group of two 5.5W motors. Expanded ToggleElevationArm() to ExtendElevationArm() and RetractElevationArm().
With the neutral flip mech disabled, we can place more motors elsewhere. This tests replacing the weak pistons with... a weak motor... will need to be geared or something. Expanding the toggle function into two different functions makes it easier to specify the extended and retracted positions, and the functions can be used during autonomous to ensure the elevationArmExtended variable is accurate.