APC Add ghosts
A drive log can now be pasted in at the top right corner and the recording will be shown on the map.
Being able to see the recorded drive next to the instructed path will make it easier to see what results in what and apply changes. This will also be a helpful tool if the line following code is ever redone, and I want a comparison between the two algorithms.