APC Merge JSON and C++ textareas
C++ is used to retrieve the path data instead of separate JSON data.
This makes it much easier and faster to edit the file, since you can copy and paste the whole thing at once instead of needing to select specific portions for the C++ code and JSON save comment.